alpha version
The changes were made to account for losing streams.

Q_gw = C_riv((S-S_ref)/Sy - H_riv)

Q_bs(t) = alpha Q_gw(t) + (1-alpha)*Q_gw(t-1)

C_riv is a flow coefficient representing the lumped effects of the
permeability of the streambed and aquifer. On the other hand, alpha in
the reservoir model is used to simulate the damping effect of the
aquifer, which depends on the aquifer diffusivity and the distance
between the HRU and stream.  Indeed, Eq 2 is similar to Eq 2:4.2.9 in
the SWAT2009 Manual.

Bank storage is ignored

The stream leakage is added to the shallow aquifer instead of the bank
and deep aquifer storages.